hmd 5

+++ 0x00-hello_world/6-size.c
@@ -1,13 +1,16 @@
 #include <stdio.h>
-/*
- * main - Entry
- * prints the size of each type
+
+/**
+ * main - Entry point
+ *
+ * Description: A C program that prints the size of each type
+ *
  * return - always 0 (Success)
  */
 int main(void)
 {
 	printf("Size of a char: %ld byte(s)\n", sizeof(char));
-	printf("Size of an int: %ld byte(s)\n", sizeof(short));
+	printf("Size of an int: %ld byte(s)\n", sizeof(int));
 	printf("Size of a long int: %ld byte(s)\n", sizeof(long int));
 	printf("Size of a long long int: %ld byte(s)\n", sizeof(long long int));
 	printf("Size of a float: %ld byte(s)\n", sizeof(float));
